assertTree doesn't rely on children order anymore
Previously assertTree always expected children to be in insertation order.
This behaviour is not guaranteed by our API.
Also tests relied on insertation order which was impossible to predict for some cases (eg. file importer).

The implemented algorithm checks all children recursively.
If all properties of all children match it is recognized as the same node.
It can deal correctly with nodes that have identical properties but different children. (eg. SNS)

May be slow for large trees and can cause StackOverflowException on deeply nested trees.